Forward template slug from empty state into the compose form
NotesPinboardEmpty's template buttons were calling
onCreate(t.slug), but ProjectNotesTab's handleCreate accepted no
arguments, so the template choice was silently dropped (every
template button produced an identical blank note).

- Extract templates to notesTemplates.ts so both empty-state and
  compose can resolve a slug to its definition
- Add an optional templateSlug to the 'creating' view kind and
  thread it through handleCreate
- NotesPinboardNew: pre-seed the tag set from the matching
  template (only on initial mount, only when not editing)
